requirements.txt
discord.py
transformers
python-dotenv
numpy
faiss-cpu
sentence-transformers
